eatgo-admin-api/src/test/java/kr/co/fastcampus/eatgo/interfaces/CategoryControllerTest.java
package kr.co.fastcampus.eatgo.interfaces;

import kr.co.fastcampus.eatgo.application.CategoryService;
import kr.co.fastcampus.eatgo.application.RegionService;
import kr.co.fastcampus.eatgo.domain.Category;
import kr.co.fastcampus.eatgo.domain.Region;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

import java.util.ArrayList;
import java.util.List;

import static org.hamcrest.Matchers.containsString;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
import static org.mockito.Mockito.*;
import static org.junit.jupiter.api.Assertions.*;

@WebMvcTest(CategoryController.class)
class CategoryControllerTest {
    @Autowired
    private MockMvc mvc;

    @MockBean
    private CategoryService categoryService;

    @Test
    public void list() throws Exception{
        List<Category> categories = new ArrayList<>();
        categories.add(Category.builder().name("Korean Food").build());
        when(categoryService.getCategories()).thenReturn(categories);
        mvc.perform(get("/categories"))
                .andExpect(status().isOk())
                .andExpect(content().string(containsString("Korean Food")))
                .andDo(print());
    }
    @Test
    public void create() throws Exception{
        // {"name":"Seoul"}
        Category category = Category.builder().name("Korean Food").build();
        when(categoryService.addCategory("Korean Food")).thenReturn(category);
        mvc.perform(post("/categories")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content("{\"name\":\"Seoul\"}")
                )
                .andExpect(status().isCreated())
                .andExpect(content().string("{}"))
                .andDo(print());

        verify(categoryService).addCategory(any());
    }
}